refactor: modularize printing function
- modularize printing function.
- make comment concise.
