+++ index.html
@@ -32,6 +32,6 @@
     <script type="module" src="src/carMake.js"></script>
     <script type="module" src="src/tryCount.js"></script>
     <script type="module" src="src/carRace.js"></script>
-
+    <script type="module" src="src/printResult.js"></script>
   </body>
 </html>
